requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-openai
langchain-core
langchain-community
langchain-text-splitters
openai
faiss-cpu
docling
python-dotenv
pydantic
pandas
pytest
pytest-asyncio
httpx
requests
streamlit

